% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroma-utils.R
\name{chroma-lightness}
\alias{chroma-lightness}
\alias{chroma_brighten}
\alias{chroma_darken}
\title{Change lightness of a color}
\usage{
chroma_brighten(color, value = 1)

chroma_darken(color, value = 1)
}
\arguments{
\item{color}{Hexadecimal string or a name for color.}

\item{value}{Value for brighten / darken.}
}
\value{
Hexadecimal string
}
\description{
Change lightness of a color
}
\examples{
chroma_brighten("#75b8d1")
chroma_darken("#75b8d1")

view_cols(c(
  chroma_darken("#75b8d1", 2),
  chroma_darken("#75b8d1"),
  "#75b8d1",
  chroma_brighten("#75b8d1"),
  chroma_brighten("#75b8d1", 2)
))
}
