% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroma-utils.R
\name{chroma_avg}
\alias{chroma_avg}
\title{Simple averaging of colors}
\usage{
chroma_avg(colors, mode = c("rgb", "hsl", "lab", "lrgb", "lch"))
}
\arguments{
\item{colors}{A vector of hexadecimal strings, minimum length 2.}

\item{mode}{Color space used for interpolation.}
}
\value{
A hexadecimal string
}
\description{
Simple averaging of colors
}
\examples{
\dontrun{
chroma_avg(colors = c("#ddd", "yellow", "red", "teal"))
}
}
