% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroma-utils.R
\name{chroma_contrast}
\alias{chroma_contrast}
\title{Contrast between two colors}
\usage{
chroma_contrast(color1, color2)
}
\arguments{
\item{color1}{Hexadecimal string or a name for color.}

\item{color2}{Hexadecimal string or a name for color.}
}
\value{
a numeric value
}
\description{
Computes the WCAG contrast ratio between two colors.
 A minimum contrast of 4.5:1 is recommended to ensure
 that text is still readable against a background color.
}
\examples{
chroma_contrast("pink", "hotpink")


# black on light grey => good
chroma_contrast("#e3e3e3", "black")

# black on dark grey => bad
chroma_contrast("#393939", "black")
}
