% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroma-utils.R
\name{chroma_distance}
\alias{chroma_distance}
\title{Distance between two colors}
\usage{
chroma_distance(color1, color2, mode = c("lab", "rgb", "hsl", "lrgb",
  "lch"))
}
\arguments{
\item{color1}{Hexadecimal string or a name for color one.}

\item{color2}{Hexadecimal string or a name for color two.}

\item{mode}{Color space used for interpolation.}
}
\value{
a numeric
}
\description{
Computes the Euclidean distance between two colors in a given color space (default is Lab).
}
\examples{
chroma_distance("#fff", "#ff0")
chroma_distance("#fff", "#ff0", "rgb")
}
