% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroma-utils.R
\name{chroma_get}
\alias{chroma_get}
\title{Returns a single channel value.}
\usage{
chroma_get(color, channel)
}
\arguments{
\item{color}{Hexadecimal string or a name for color.}

\item{channel}{Channel desired, for example \code{"hsl.h"} for hue parameter in hsl color space.}
}
\value{
a single numeric value
}
\description{
Returns a single channel value.
}
\examples{
chroma_get("#1D9A6C", "hsl.h")
chroma_get("hotpink", "hsl.h")

# vectorized over color
chroma_get(c("hotpink", "steelblue"), "rgb.g")
}
