% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroma-utils.R
\name{chroma_mix}
\alias{chroma_mix}
\title{Mix two colors}
\usage{
chroma_mix(color1, color2, ratio = 0.5, mode = c("rgb", "hsl", "lab",
  "lrgb", "lch"))
}
\arguments{
\item{color1}{Hexadecimal string or a name for color one.}

\item{color2}{Hexadecimal string or a name for color two.}

\item{ratio}{A value between 0 and 1.}

\item{mode}{Color space used for interpolation.}
}
\value{
A hexadecimal string
}
\description{
Mix two colors
}
\examples{
chroma_mix("red", "blue")

chroma_mix("red", "blue", ratio = 0.75)

chroma_mix("red", "blue", mode = "lab")
}
