% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroma-utils.R
\name{chroma_random}
\alias{chroma_random}
\title{Random colors}
\usage{
chroma_random(n = 1L)
}
\arguments{
\item{n}{Number of colors desired.}
}
\value{
A vector of hexadecimal string(s).
}
\description{
Creates a random color by generating a random hexadecimal string.
}
\examples{
chroma_random()
chroma_random(10)
}
