% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorscale-addin.R
\name{color_scale}
\alias{color_scale}
\title{Color Scale Shiny Gadget}
\usage{
color_scale(color = "#1D9A6C")
}
\arguments{
\item{color}{Hexadecimal string or a name for color.}
}
\description{
Interactively create a color palette from a unique color
}
\examples{
\dontrun{

if (interactive()) {

# Launch the gadget with :
color_scale()

}

}
}
