% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{colors_rect}
\alias{colors_rect}
\title{Create rectangles in HTML to display colors}
\usage{
colors_rect(colors, width = "100\%", height = "45px")
}
\arguments{
\item{colors}{A vector of colors to visualize}

\item{width}{The width of the HTML tag, e.g. '400px', or '100\%'.}

\item{height}{The height of the HTML tag, e.g. '400px', or '100\%'.}
}
\value{
An HTML tag
}
\description{
Create rectangles in HTML to display colors
}
\examples{
\dontrun{

if (interactive()) {

library(shiny)
library(colorscale)

ui <- fluidPage(
  tags$h2("Visualizing colors in Shiny"),
  colors_rect(colors = c(
    get_dark_cols("#1D9A6C"),
    "#1D9A6C",
    get_light_cols("#1D9A6C")
  ))
)

server <- function(input, output, session) {

}

shinyApp(ui, server)

}

}
}
