% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{continuous-palettes}
\alias{continuous-palettes}
\alias{chroma_continuous_pal}
\alias{bezier_continuous_pal}
\title{Continuous color palettes}
\usage{
chroma_continuous_pal(
  colors = c("yellow", "navy"),
  mode = "lab",
  padding = NULL,
  correct_lightness = TRUE
)

bezier_continuous_pal(colors = c("yellow", "red", "black"))
}
\arguments{
\item{colors}{Vector of hexadecimal strings or names for colors.}

\item{mode}{Color space used for interpolation.}

\item{padding}{Reduces the color range by cutting of a fraction of the gradient on both sides.
If you pass a single number, the same padding will be applied to both ends.}

\item{correct_lightness}{Makes sure the lightness range is spread evenly across a color scale.}
}
\value{
a function
}
\description{
Continuous color palettes
}
\examples{
x <- seq(0, 1, length.out = 25)
view_cols(chroma_continuous_pal()(x))

# change mode
view_cols(chroma_continuous_pal(mode = "lch")(x))


# Bezier interpolation
view_cols(bezier_continuous_pal()(x))

# Color Brewer :
view_cols(chroma_continuous_pal("RdYlBu")(x))
# add some padding :
view_cols(chroma_continuous_pal("RdYlBu", padding = 0.15)(x))

}
