% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-colorscale.R
\name{create_single_scale}
\alias{create_single_scale}
\title{Create a color scale from a single color}
\usage{
create_single_scale(
  color,
  n_light = 6,
  n_dark = 4,
  lightness = 0.8,
  darkness = 0.5,
  rotate_light = 67,
  rotate_dark = -51,
  saturation_light = 0.2,
  saturation_dark = 0.14
)
}
\arguments{
\item{color}{Hexadecimal string or a name of a color.}

\item{n_light}{Number of light colors to generate.}

\item{n_dark}{Number of dark colors to generate.}

\item{lightness}{Percentage of lightness (\code{[0, 1]}).}

\item{darkness}{Percentage of darkness (\code{[0, 1]}).}

\item{rotate_light}{Rotate coefficient in HSL space for light colors (\code{[-360, 360]}).}

\item{rotate_dark}{Rotate coefficient in HSL space for dark colors (\code{[-360, 360]}).}

\item{saturation_light}{Saturation percentage in HSL space for light colors (\code{[-1, 1]}).}

\item{saturation_dark}{Saturation percentage in HSL space for dark colors (\code{[-1, 1]}).}
}
\value{
a vector of \code{n_light + n_dark + 1} colors.
}
\description{
Create a color scale from a single color
}
\examples{
create_single_scale(color = "#1D9A6C")

# display in viewer
view_cols(create_single_scale(color = "#1D9A6C"))
}
