% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorscale-addin.R
\name{one_color_scale}
\alias{one_color_scale}
\title{Color Scale Shiny Gadget}
\usage{
one_color_scale(
  color = "#1D9A6C",
  viewer = getOption(x = "colorscale.viewer", default = "pane")
)
}
\arguments{
\item{color}{Hexadecimal string or a name for color.}

\item{viewer}{Where to display the gadget: \code{"dialog"},
\code{"pane"} or \code{"browser"} (see \code{\link[shiny]{viewer}}).}
}
\description{
Interactively create a color palette from a unique color
}
\examples{
\dontrun{

if (interactive()) {

# Launch the gadget with :
one_color_scale()

}

}
}
