% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_cols.R
\name{view-colors}
\alias{view-colors}
\alias{view_cols}
\alias{view_pal}
\alias{view_gradient}
\title{Display colors in the viewer}
\usage{
view_cols(colors, width = 80, height = 80)

view_pal(pal, height = 80)

view_gradient(colors, height = 80)
}
\arguments{
\item{colors}{A vector of colors}

\item{width}{Width of each box.}

\item{height}{Height of each box}

\item{pal}{A function for generating a palette.}
}
\description{
Display colors in the viewer
}
\examples{

### View colors
# Scale of greys
view_cols(c("#ffffff", "#e3e3e3", "#c6c6c6", "#aaaaaa",
            "#8e8e8e", "#717171", "#555555", "#393939",
            "#1c1c1c", "#000000"))


# 50 first colors from colors() (converted to HEX first)
col2hex <- function(col, alpha) # function from @davidorme
  rgb(t(col2rgb(col)), alpha=alpha, maxColorValue=255)


view_cols(col2hex(colors()[1:50]))


### View palettes

# View one palette
view_pal(pal = chroma_continuous_pal())
view_pal(pal = chroma_continuous_pal(mode = "lch"))

# View several palettes
view_pal(pal = list(
  "default (mode lab)" = chroma_continuous_pal(),
  "mode lch" = chroma_continuous_pal(mode = "lch"),
  "mode rgb" = chroma_continuous_pal(mode = "rgb")
))

# View scales palettes
library(scales)
view_pal(pal = list(
  "seq_gradient_pal" = seq_gradient_pal(),
  "gradient_n_pal" = gradient_n_pal(c("yellow", "navy"))
))


### View a gradient

x <- seq(0, 1, length.out = 100)
view_gradient(chroma_continuous_pal()(x))

# with a list
view_gradient(list(
  "chroma" = chroma_continuous_pal()(x),
  "bezier" = bezier_continuous_pal()(x)
))
}
