% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_cols.R
\name{view_cols}
\alias{view_cols}
\title{Display colors in the viewer}
\usage{
view_cols(colors, width = 80, height = 80)
}
\arguments{
\item{colors}{A vector of colors}

\item{width}{Width of each box.}

\item{height}{Height of each box}
}
\description{
Display colors in the viewer
}
\examples{
\dontrun{

# Scale of greys
view_cols(c("#ffffff", "#e3e3e3", "#c6c6c6", "#aaaaaa",
            "#8e8e8e", "#717171", "#555555", "#393939",
            "#1c1c1c", "#000000"))


# 50 first colors from colors() (converted to HEX first)
col2hex <- function(col, alpha) # function from @davidorme
  rgb(t(col2rgb(col)), alpha=alpha, maxColorValue=255)


view_cols(col2hex(colors()[1:50]))

}
}
