% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroma-class.R
\name{chroma}
\alias{chroma}
\title{Chroma API}
\description{
Manipulate colors with 'chroma.js' library.
See \url{https://gka.github.io/chroma.js/}.
}
\note{
Dont initialize chroma instance with a color.

Dont initialize chroma instance with a color.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{chroma$new()}}
\item \href{#method-chroma.hsl}{\code{chroma$chroma.hsl()}}
\item \href{#method-chroma.hsv}{\code{chroma$chroma.hsv()}}
\item \href{#method-chroma.lch}{\code{chroma$chroma.lch()}}
\item \href{#method-chroma.gl}{\code{chroma$chroma.gl()}}
\item \href{#method-mix}{\code{chroma$mix()}}
\item \href{#method-average}{\code{chroma$average()}}
\item \href{#method-blend}{\code{chroma$blend()}}
\item \href{#method-random}{\code{chroma$random()}}
\item \href{#method-contrast}{\code{chroma$contrast()}}
\item \href{#method-distance}{\code{chroma$distance()}}
\item \href{#method-alpha}{\code{chroma$alpha()}}
\item \href{#method-brighten}{\code{chroma$brighten()}}
\item \href{#method-darken}{\code{chroma$darken()}}
\item \href{#method-saturate}{\code{chroma$saturate()}}
\item \href{#method-desaturate}{\code{chroma$desaturate()}}
\item \href{#method-set}{\code{chroma$set()}}
\item \href{#method-get}{\code{chroma$get()}}
\item \href{#method-luminance}{\code{chroma$luminance()}}
\item \href{#method-hex}{\code{chroma$hex()}}
\item \href{#method-name}{\code{chroma$name()}}
\item \href{#method-css}{\code{chroma$css()}}
\item \href{#method-rgb}{\code{chroma$rgb()}}
\item \href{#method-rgba}{\code{chroma$rgba()}}
\item \href{#method-hsl}{\code{chroma$hsl()}}
\item \href{#method-hsv}{\code{chroma$hsv()}}
\item \href{#method-hsi}{\code{chroma$hsi()}}
\item \href{#method-lab}{\code{chroma$lab()}}
\item \href{#method-lch}{\code{chroma$lch()}}
\item \href{#method-hcl}{\code{chroma$hcl()}}
\item \href{#method-gl}{\code{chroma$gl()}}
\item \href{#method-print}{\code{chroma$print()}}
\item \href{#method-eval}{\code{chroma$eval()}}
\item \href{#method-clone}{\code{chroma$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new chroma instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$new(color = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{color}}{Name, hex code or a vector of length 3 (rgb) of a color.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `chroma` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chroma.hsl"></a>}}
\subsection{Method \code{chroma.hsl()}}{
Create a new chroma instance with HSL color.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$chroma.hsl(hue, saturation, lightness)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hue}}{Hue parameter.}

\item{\code{saturation}}{Aaturation parameter.}

\item{\code{lightness}}{Lightness parameter.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `chroma` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chroma.hsv"></a>}}
\subsection{Method \code{chroma.hsv()}}{
Create a new chroma instance with HSV color.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$chroma.hsv(hue, saturation, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hue}}{Hue parameter.}

\item{\code{saturation}}{Saturation parameter.}

\item{\code{value}}{Value parameter.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `chroma` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chroma.lch"></a>}}
\subsection{Method \code{chroma.lch()}}{
Create a new chroma instance with LCH color.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$chroma.lch(Lightness, chroma, hue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Lightness}}{Lightness parameter.}

\item{\code{chroma}}{Chroma parameter.}

\item{\code{hue}}{Hue parameter.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `chroma` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chroma.gl"></a>}}
\subsection{Method \code{chroma.gl()}}{
Create a new chroma instance with RGB color.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$chroma.gl(red, green, blue, alpha = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{red}}{Red parameter.}

\item{\code{green}}{Green parameter.}

\item{\code{blue}}{Blue parameter.}

\item{\code{alpha}}{Alpha parameter.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `chroma` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mix"></a>}}
\subsection{Method \code{mix()}}{
Mixes two colors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$mix(color2, ratio = 0.5, mode = c("rgb", "hsl", "lab", "lrgb", "lch"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{color2}}{Name of a second color.}

\item{\code{ratio}}{A value between 0 and 1.}

\item{\code{mode}}{Color space used for interpolation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `chroma` object modified.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-average"></a>}}
\subsection{Method \code{average()}}{
Average between colors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$average(colors, mode = c("rgb", "hsl", "lab", "lrgb", "lch"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{colors}}{Color(s).}

\item{\code{mode}}{Color space used for interpolation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `chroma` object modified.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-blend"></a>}}
\subsection{Method \code{blend()}}{
Blend two colors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$blend(
  color2,
  type = c("multiply", "darken", "lighten", "screen", "overlay", "burn", "dodge")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{color2}}{Name of a second color.}

\item{\code{blend}}{Blend mode.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `chroma` object modified.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random"></a>}}
\subsection{Method \code{random()}}{
Get a random color
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$random()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A `chroma` object modified.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-contrast"></a>}}
\subsection{Method \code{contrast()}}{
Contrast between two colors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$contrast(color1, color2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{color1}}{Name of first color.}

\item{\code{color2}}{Name of second color.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `chroma` object modified.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distance"></a>}}
\subsection{Method \code{distance()}}{
Distancec between two colors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$distance(color1, color2, mode = c("lab", "rgb", "hsl", "lrgb", "lch"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{color1}}{Name of first color.}

\item{\code{color2}}{Name of second color.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `chroma` object modified.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-alpha"></a>}}
\subsection{Method \code{alpha()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$alpha(value = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-brighten"></a>}}
\subsection{Method \code{brighten()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$brighten(value = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-darken"></a>}}
\subsection{Method \code{darken()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$darken(value = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saturate"></a>}}
\subsection{Method \code{saturate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$saturate(value = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-desaturate"></a>}}
\subsection{Method \code{desaturate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$desaturate(value = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$set(channel, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$get(channel)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-luminance"></a>}}
\subsection{Method \code{luminance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$luminance(lum = NULL, mode = "rgb")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hex"></a>}}
\subsection{Method \code{hex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$hex()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-name"></a>}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-css"></a>}}
\subsection{Method \code{css()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$css()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rgb"></a>}}
\subsection{Method \code{rgb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$rgb(round = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rgba"></a>}}
\subsection{Method \code{rgba()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$rgba(round = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hsl"></a>}}
\subsection{Method \code{hsl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$hsl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hsv"></a>}}
\subsection{Method \code{hsv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$hsv()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hsi"></a>}}
\subsection{Method \code{hsi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$hsi()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lab"></a>}}
\subsection{Method \code{lab()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$lab()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lch"></a>}}
\subsection{Method \code{lch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$lch()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hcl"></a>}}
\subsection{Method \code{hcl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$hcl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gl"></a>}}
\subsection{Method \code{gl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$gl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval"></a>}}
\subsection{Method \code{eval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$eval(type_convert = TRUE, split_char = ",")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
