% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroma-class.R
\docType{data}
\name{chroma_scale}
\alias{chroma_scale}
\title{Chroma Scale API}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
chroma_scale
}
\description{
Create a color scale with chroma.js
}
\keyword{datasets}
