% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroma-class.R
\name{chroma_scale}
\alias{chroma_scale}
\title{Chroma Scale API}
\description{
Create a color scale with chroma.js
}
\examples{
# Palette from white to black
cs <- chroma_scale$new()
cs$scale()
cs$colors(10)
(res <- cs$eval())
view_cols(res)


# Specify from & to colors
cs <- chroma_scale$new()
cs$scale(colors = c("yellow", "#008ae5"))
cs$colors(10)
(res <- cs$eval())
view_cols(res)


# With three colors
cs <- chroma_scale$new()
cs$scale(colors = c("yellow", "red", "black"))
cs$colors(20)
(res <- cs$eval())
view_cols(res)


# Default color space for interpolation is RGB
cs <- chroma_scale$new()
cs$scale(colors = c("yellow", "navy"))
cs$colors(20)
(res <- cs$eval())
view_cols(res)

# Change with mode
cs$mode("lab")
(res <- cs$eval())
view_cols(res)



# Correct lightness example
cs <- chroma_scale$new()
cs$scale(c("black","red","yellow","white"))
cs$colors(n = 16)

# without
(res <- cs$eval())
view_cols(res)

# with correction
cs$correctLightness()
(res <- cs$eval())
view_cols(res)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{chroma_scale$new()}}
\item \href{#method-scale}{\code{chroma_scale$scale()}}
\item \href{#method-bezier}{\code{chroma_scale$bezier()}}
\item \href{#method-mode}{\code{chroma_scale$mode()}}
\item \href{#method-correctLightness}{\code{chroma_scale$correctLightness()}}
\item \href{#method-colors}{\code{chroma_scale$colors()}}
\item \href{#method-domain}{\code{chroma_scale$domain()}}
\item \href{#method-gamma}{\code{chroma_scale$gamma()}}
\item \href{#method-padding}{\code{chroma_scale$padding()}}
\item \href{#method-print}{\code{chroma_scale$print()}}
\item \href{#method-eval}{\code{chroma_scale$eval()}}
\item \href{#method-clone}{\code{chroma_scale$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma_scale$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scale"></a>}}
\subsection{Method \code{scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma_scale$scale(colors)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bezier"></a>}}
\subsection{Method \code{bezier()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma_scale$bezier(colors)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mode"></a>}}
\subsection{Method \code{mode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma_scale$mode(mode = c("rgb", "lab", "lrgb", "hsl", "lch"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-correctLightness"></a>}}
\subsection{Method \code{correctLightness()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma_scale$correctLightness()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-colors"></a>}}
\subsection{Method \code{colors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma_scale$colors(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-domain"></a>}}
\subsection{Method \code{domain()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma_scale$domain(domain)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gamma"></a>}}
\subsection{Method \code{gamma()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma_scale$gamma(gamma)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-padding"></a>}}
\subsection{Method \code{padding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma_scale$padding(padding)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma_scale$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval"></a>}}
\subsection{Method \code{eval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma_scale$eval(type_convert = TRUE, split_char = ",")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{chroma_scale$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
