% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-colorscale.R
\name{get_dark_cols}
\alias{get_dark_cols}
\title{Get n darkest colors from a single one}
\usage{
get_dark_cols(color, n = 4, darkness = 0.5, rotate = -51,
  saturation = 0.14)
}
\arguments{
\item{color}{Hexadecimal string or a name for color.}

\item{n}{Number of desired colors.}

\item{darkness}{Percentage of darkness (\code{[0, 1]}).}

\item{rotate}{Rotate coefficient in HSL space (\code{[-360, 360]}).}

\item{saturation}{Saturation percentage in HSL space (\code{[-1, 1]}).}
}
\value{
a vector of \code{n} colors.
}
\description{
Get n darkest colors from a single one
}
\examples{
get_dark_cols("#1D9A6C")

# display in viewer
view_cols(get_dark_cols(color = "#1D9A6C"))
}
