% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-colorscale.R
\name{get_light_cols}
\alias{get_light_cols}
\title{Get n lightest colors from a single one}
\usage{
get_light_cols(color, n = 6, lightness = 0.8, rotate = 67,
  saturation = 0.2)
}
\arguments{
\item{color}{Hexadecimal string or a name for color.}

\item{n}{Number of desired colors.}

\item{lightness}{Percentage of lightness.}

\item{rotate}{Rotate coefficient in HSL space.}

\item{saturation}{Saturation percentage in HSL space}
}
\value{
a vector of \code{n} colors
}
\description{
Get n lightest colors from a single one
}
\examples{
get_light_cols("#1D9A6C")
}
